movefree_indoor/requirements.txt
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0
pillow>=10.0.0
matplotlib>=3.7.0
pandas>=2.0.0
seaborn>=0.12.0
pyyaml>=6.0
numpy>=1.24.0
pyttsx3>=2.90
SpeechRecognition>=3.10.0
pyaudio>=0.2.13
flask>=2.3.0
flask-cors>=4.0.0
websockets>=11.0
tqdm>=4.65.0
